[Security Solution] remove socTrendsEnabled feature flag and related code
