+++ x-pack/solutions/security/plugins/security_solution/public/common/mock/global_state.ts
@@ -291,7 +291,7 @@ export const mockGlobalState: State = {
         from: '2020-07-07T08:20:18.966Z',
         to: '2020-07-08T08:20:18.966Z',
       },
-      linkTo: [InputsModelId.timeline, InputsModelId.socTrends],
+      linkTo: [InputsModelId.timeline],
       queries: [],
       policy: { kind: DEFAULT_INTERVAL_TYPE, duration: DEFAULT_INTERVAL_VALUE },
       query: {
@@ -308,7 +308,7 @@ export const mockGlobalState: State = {
         from: '2020-07-07T08:20:18.966Z',
         to: '2020-07-08T08:20:18.966Z',
       },
-      linkTo: [InputsModelId.global, InputsModelId.socTrends],
+      linkTo: [InputsModelId.global],
       queries: [],
       policy: { kind: DEFAULT_INTERVAL_TYPE, duration: DEFAULT_INTERVAL_VALUE },
       query: {
@@ -316,17 +316,6 @@ export const mockGlobalState: State = {
         language: 'kuery',
       },
       filters: [],
-    },
-    socTrends: {
-      timerange: {
-        kind: 'relative',
-        fromStr: DEFAULT_FROM,
-        toStr: DEFAULT_TO,
-        from: '2020-07-06T08:20:18.966Z',
-        to: '2020-07-07T08:20:18.966Z',
-      },
-      linkTo: [InputsModelId.global, InputsModelId.timeline],
-      policy: { kind: DEFAULT_INTERVAL_TYPE, duration: DEFAULT_INTERVAL_VALUE },
     },
     valueReport: {
       timerange: {
