+++ x-pack/solutions/security/plugins/security_solution/public/common/store/inputs/constants.ts
@@ -8,7 +8,6 @@
 export enum InputsModelId {
   global = 'global',
   timeline = 'timeline',
-  socTrends = 'socTrends',
   valueReport = 'valueReport',
 }
