+++ x-pack/solutions/security/plugins/security_solution/public/common/store/inputs/helpers.test.ts
@@ -9,7 +9,7 @@ import { cloneDeep } from 'lodash/fp';
 
 import { mockGlobalState } from '../../mock';
 
-import type { UpdateQueryParams, SetIsInspectedParams } from './helpers';
+import type { SetIsInspectedParams, UpdateQueryParams } from './helpers';
 import {
   addInputLink,
   deleteOneQuery,
@@ -24,7 +24,6 @@ import { InputsModelId } from './constants';
 
 let state = mockGlobalState.inputs;
 
-// TODO: remove socTrends check when socTrendsEnabled feature flag removed
 const emptyLinkToState: InputsModel = {
   ...state,
   global: {
@@ -35,33 +34,6 @@ const emptyLinkToState: InputsModel = {
     ...state.timeline,
     linkTo: [],
   },
-  ...(state.socTrends
-    ? {
-        socTrends: {
-          ...state.socTrends,
-          linkTo: [],
-        },
-      }
-    : {}),
-};
-const socTrendsGlobalLinkToState: InputsModel = {
-  ...state,
-  global: {
-    ...state.global,
-    linkTo: [InputsModelId.socTrends],
-  },
-  timeline: {
-    ...state.timeline,
-    linkTo: [],
-  },
-  ...(state.socTrends
-    ? {
-        socTrends: {
-          ...state.socTrends,
-          linkTo: [InputsModelId.global],
-        },
-      }
-    : {}),
 };
 const timelineGlobalLinkToState: InputsModel = {
   ...state,
@@ -73,14 +45,6 @@ const timelineGlobalLinkToState: InputsModel = {
     ...state.timeline,
     linkTo: [InputsModelId.global],
   },
-  ...(state.socTrends
-    ? {
-        socTrends: {
-          ...state.socTrends,
-          linkTo: [],
-        },
-      }
-    : {}),
 };
 const allLinkToState: InputsModel = state;
 
@@ -100,8 +64,8 @@ describe('Inputs', () => {
   });
 
   describe('#updateInputTimerange', () => {
-    describe('when timeline, global, and socTrends are lock', () => {
-      test('timeline should be identical to global & socTrends should be previous time range when global change', () => {
+    describe('when timeline and global are lock', () => {
+      test('timeline should be identical to global', () => {
         const newTimerange: TimeRange = {
           kind: 'relative',
           fromStr: 'now-48h',
@@ -115,14 +79,9 @@ describe('Inputs', () => {
           allLinkToState
         );
         expect(newState.timeline.timerange).toEqual(newState.global.timerange);
-        expect(newState.socTrends?.timerange).toEqual({
-          kind: 'absolute',
-          from: '2020-07-04T08:00:00.000Z',
-          to: '2020-07-06T08:00:00.000Z',
-        });
       });
 
-      test('global should be identical to timeline & socTrends should be previous time range when timeline change', () => {
+      test('global should be identical to timeline', () => {
         const newTimerange: TimeRange = {
           kind: 'relative',
           fromStr: 'now-68h',
@@ -136,38 +95,11 @@ describe('Inputs', () => {
           allLinkToState
         );
         expect(newState.timeline.timerange).toEqual(newState.global.timerange);
-        expect(newState.socTrends?.timerange).toEqual({
-          kind: 'absolute',
-          from: '2020-07-03T02:00:00.000Z',
-          to: '2020-07-05T22:00:00.000Z',
-        });
-      });
-
-      test('global and timeline should be future time range when socTrends change', () => {
-        const newTimerange: TimeRange = {
-          kind: 'relative',
-          fromStr: 'now-68h',
-          toStr: 'NOTnow',
-          from: '2020-07-01T00:00:00.000Z',
-          to: '2020-07-02T00:00:00.000Z',
-        };
-        const newState: InputsModel = updateInputTimerange(
-          InputsModelId.socTrends,
-          newTimerange,
-          allLinkToState
-        );
-        expect(newState.global.timerange).toEqual({
-          kind: 'absolute',
-          from: '2020-07-02T00:00:00.000Z',
-          to: '2020-07-03T00:00:00.000Z',
-        });
-        expect(newState.global.timerange).toEqual(newState.timeline.timerange);
-        expect(newState.socTrends?.timerange).toEqual(newTimerange);
       });
     });
 
-    describe('when timeline, global, and socTrends are NOT lock', () => {
-      test('timeline and socTrends should remain unchanged when global change', () => {
+    describe('when timeline and global are NOT lock', () => {
+      test('timeline should remain unchanged when global change', () => {
         const newTimerange: TimeRange = {
           kind: 'relative',
           fromStr: 'now-48h',
@@ -181,11 +113,10 @@ describe('Inputs', () => {
           emptyLinkToState
         );
         expect(newState.timeline.timerange).toEqual(emptyLinkToState.timeline.timerange);
-        expect(newState.socTrends?.timerange).toEqual(emptyLinkToState.socTrends?.timerange);
         expect(newState.global.timerange).toEqual(newTimerange);
       });
 
-      test('global and socTrends should remain unchanged when timeline change', () => {
+      test('global should remain unchanged when timeline change', () => {
         const newTimerange: TimeRange = {
           kind: 'relative',
           fromStr: 'now-68h',
@@ -199,152 +130,7 @@ describe('Inputs', () => {
           emptyLinkToState
         );
         expect(newState.timeline.timerange).toEqual(newTimerange);
-        expect(newState.socTrends?.timerange).toEqual(emptyLinkToState.socTrends?.timerange);
         expect(newState.global.timerange).toEqual(emptyLinkToState.global.timerange);
-      });
-
-      test('timeline and global should remain unchanged when socTrends change', () => {
-        const newTimerange: TimeRange = {
-          kind: 'relative',
-          fromStr: 'now-48h',
-          toStr: 'now',
-          from: '2020-07-06T08:00:00.000Z',
-          to: '2020-07-08T08:00:00.000Z',
-        };
-        const newState: InputsModel = updateInputTimerange(
-          InputsModelId.socTrends,
-          newTimerange,
-          emptyLinkToState
-        );
-        expect(newState.timeline.timerange).toEqual(emptyLinkToState.timeline.timerange);
-        expect(newState.global.timerange).toEqual(emptyLinkToState.global.timerange);
-        expect(newState.socTrends?.timerange).toEqual(newTimerange);
-      });
-    });
-
-    describe('when timeline and global only are lock', () => {
-      test('timeline should be identical to global & socTrends should remain unchanged when global change', () => {
-        const newTimerange: TimeRange = {
-          kind: 'relative',
-          fromStr: 'now-48h',
-          toStr: 'now',
-          from: '2020-07-06T08:00:00.000Z',
-          to: '2020-07-08T08:00:00.000Z',
-        };
-        const newState: InputsModel = updateInputTimerange(
-          InputsModelId.global,
-          newTimerange,
-          timelineGlobalLinkToState
-        );
-        expect(newState.timeline.timerange).toEqual(newState.global.timerange);
-        expect(newState.socTrends?.timerange).toEqual(
-          timelineGlobalLinkToState.socTrends?.timerange
-        );
-      });
-
-      test('global should be identical to timeline & socTrends should be previous time range when timeline change', () => {
-        const newTimerange: TimeRange = {
-          kind: 'relative',
-          fromStr: 'now-68h',
-          toStr: 'NOTnow',
-          from: '2020-07-05T22:00:00.000Z',
-          to: '2020-07-08T18:00:00.000Z',
-        };
-        const newState: InputsModel = updateInputTimerange(
-          InputsModelId.timeline,
-          newTimerange,
-          timelineGlobalLinkToState
-        );
-        expect(newState.timeline.timerange).toEqual(newState.global.timerange);
-        expect(newState.socTrends?.timerange).toEqual(
-          timelineGlobalLinkToState.socTrends?.timerange
-        );
-      });
-
-      test('global and timeline should remain unchanged when socTrends change', () => {
-        const newTimerange: TimeRange = {
-          kind: 'relative',
-          fromStr: 'now-68h',
-          toStr: 'NOTnow',
-          from: '2020-07-01T00:00:00.000Z',
-          to: '2020-07-02T00:00:00.000Z',
-        };
-        const newState: InputsModel = updateInputTimerange(
-          InputsModelId.socTrends,
-          newTimerange,
-          timelineGlobalLinkToState
-        );
-
-        expect(newState.global.timerange).toEqual(timelineGlobalLinkToState.global.timerange);
-        expect(newState.timeline.timerange).toEqual(timelineGlobalLinkToState.timeline.timerange);
-        expect(newState.socTrends?.timerange).toEqual(newTimerange);
-      });
-    });
-
-    describe('when socTrends and global only are lock', () => {
-      test('timeline remain unchanged & socTrends should be previous time range when global change', () => {
-        const newTimerange: TimeRange = {
-          kind: 'relative',
-          fromStr: 'now-48h',
-          toStr: 'now',
-          from: '2020-07-06T08:00:00.000Z',
-          to: '2020-07-08T08:00:00.000Z',
-        };
-        const newState: InputsModel = updateInputTimerange(
-          InputsModelId.global,
-          newTimerange,
-          socTrendsGlobalLinkToState
-        );
-        expect(newState.timeline.timerange).toEqual(socTrendsGlobalLinkToState.timeline.timerange);
-        expect(newState.global.timerange).toEqual(newTimerange);
-        expect(newState.socTrends?.timerange).toEqual({
-          kind: 'absolute',
-          from: '2020-07-04T08:00:00.000Z',
-          to: '2020-07-06T08:00:00.000Z',
-        });
-      });
-
-      test('global & socTrends should remain unchanged when timeline change', () => {
-        const newTimerange: TimeRange = {
-          kind: 'relative',
-          fromStr: 'now-68h',
-          toStr: 'NOTnow',
-          from: '2020-07-05T22:00:00.000Z',
-          to: '2020-07-08T18:00:00.000Z',
-        };
-        const newState: InputsModel = updateInputTimerange(
-          InputsModelId.timeline,
-          newTimerange,
-          socTrendsGlobalLinkToState
-        );
-        expect(newState.timeline.timerange).toEqual(newTimerange);
-        expect(newState.global.timerange).toEqual(socTrendsGlobalLinkToState.global.timerange);
-        expect(newState.socTrends?.timerange).toEqual(
-          socTrendsGlobalLinkToState.socTrends?.timerange
-        );
-      });
-
-      test('global should update to future time and timeline should remain unchanged when socTrends change', () => {
-        const newTimerange: TimeRange = {
-          kind: 'relative',
-          fromStr: 'now-68h',
-          toStr: 'NOTnow',
-          from: '2020-07-01T00:00:00.000Z',
-          to: '2020-07-02T00:00:00.000Z',
-        };
-        const newState: InputsModel = updateInputTimerange(
-          InputsModelId.socTrends,
-          newTimerange,
-          socTrendsGlobalLinkToState
-        );
-
-        expect(newState.timeline.timerange).toEqual(socTrendsGlobalLinkToState.timeline.timerange);
-        expect(newState.global.timerange).toEqual({
-          kind: 'absolute',
-          from: '2020-07-02T00:00:00.000Z',
-          to: '2020-07-03T00:00:00.000Z',
-        });
-        expect(newState.socTrends?.timerange).toEqual(newTimerange);
       });
     });
   });
@@ -456,30 +242,19 @@ describe('Inputs', () => {
 
   describe('#addInputLink', () => {
     describe(`does not allow bad values`, () => {
-      test('More than 2 linkToIds passed', () => {
-        expect(() =>
-          addInputLink(
-            [InputsModelId.global, InputsModelId.timeline, InputsModelId.socTrends],
-            state
-          )
-        ).toThrow('Only link 2 input states at a time');
-      });
       test('Less than 2 linkToIds passed', () => {
         expect(() => addInputLink([InputsModelId.global], state)).toThrow(
           'Only link 2 input states at a time'
         );
       });
+
       test('Identical linkToIds passed', () => {
         expect(() => addInputLink([InputsModelId.global, InputsModelId.global], state)).toThrow(
           'Input linkTo cannot link to itself'
         );
       });
-      test('Do not link timeline and socTrends', () => {
-        expect(() =>
-          addInputLink([InputsModelId.timeline, InputsModelId.socTrends], state)
-        ).toThrow('Do not link socTrends to timeline. Only link socTrends to global');
-      });
     });
+
     describe(`linkToIds === ["global", "timeline"]`, () => {
       test('no inputs linked, add timeline to global, add global to timeline', () => {
         const newState: InputsModel = addInputLink(
@@ -488,23 +263,8 @@ describe('Inputs', () => {
         );
         expect(newState.global.linkTo).toEqual([InputsModelId.timeline]);
         expect(newState.timeline.linkTo).toEqual([InputsModelId.global]);
-        expect(newState.socTrends?.linkTo).toEqual([]);
       });
-      test('socTends and global linked, add timeline and socTrends to global, add global and socTrends to timeline', () => {
-        const newState: InputsModel = addInputLink(
-          [InputsModelId.global, InputsModelId.timeline],
-          socTrendsGlobalLinkToState
-        );
-        expect(newState.global.linkTo.sort()).toEqual(
-          [InputsModelId.socTrends, InputsModelId.timeline].sort()
-        );
-        expect(newState.timeline.linkTo.sort()).toEqual(
-          [InputsModelId.global, InputsModelId.socTrends].sort()
-        );
-        expect(newState.socTrends?.linkTo.sort()).toEqual(
-          [InputsModelId.global, InputsModelId.timeline].sort()
-        );
-      });
+
       test('timeline and global linked, do not update state', () => {
         const newState: InputsModel = addInputLink(
           [InputsModelId.global, InputsModelId.timeline],
@@ -516,99 +276,34 @@ describe('Inputs', () => {
         expect(newState.timeline.linkTo.sort()).toEqual(
           timelineGlobalLinkToState.timeline.linkTo.sort()
         );
-        expect(newState.socTrends?.linkTo.sort()).toEqual(
-          timelineGlobalLinkToState.socTrends?.linkTo.sort()
-        );
       });
-      test('socTends, timeline, and global linked, do not update state', () => {
+
+      test('timeline, and global linked, do not update state', () => {
         const newState: InputsModel = addInputLink(
           [InputsModelId.global, InputsModelId.timeline],
           allLinkToState
         );
         expect(newState.global.linkTo.sort()).toEqual(allLinkToState.global.linkTo.sort());
         expect(newState.timeline.linkTo.sort()).toEqual(allLinkToState.timeline.linkTo.sort());
-        expect(newState.socTrends?.linkTo.sort()).toEqual(allLinkToState.socTrends?.linkTo.sort());
-      });
-    });
-    describe(`linkToIds === ["global", "socTrends"]`, () => {
-      test('no inputs linked, add socTrends to global, add global to socTrends', () => {
-        const newState: InputsModel = addInputLink(
-          [InputsModelId.global, InputsModelId.socTrends],
-          emptyLinkToState
-        );
-        expect(newState.global.linkTo).toEqual([InputsModelId.socTrends]);
-        expect(newState.timeline.linkTo).toEqual([]);
-        expect(newState.socTrends?.linkTo).toEqual([InputsModelId.global]);
-      });
-      test('timeline and global linked, add timeline and socTrends to global, add global and timeline to socTrends', () => {
-        const newState: InputsModel = addInputLink(
-          [InputsModelId.global, InputsModelId.socTrends],
-          timelineGlobalLinkToState
-        );
-        expect(newState.global.linkTo.sort()).toEqual(
-          [InputsModelId.socTrends, InputsModelId.timeline].sort()
-        );
-        expect(newState.timeline.linkTo.sort()).toEqual(
-          [InputsModelId.global, InputsModelId.socTrends].sort()
-        );
-        expect(newState.socTrends?.linkTo.sort()).toEqual(
-          [InputsModelId.global, InputsModelId.timeline].sort()
-        );
-      });
-      test('socTends and global linked, do not update state', () => {
-        const newState: InputsModel = addInputLink(
-          [InputsModelId.global, InputsModelId.socTrends],
-          socTrendsGlobalLinkToState
-        );
-        expect(newState.global.linkTo.sort()).toEqual(
-          socTrendsGlobalLinkToState.global.linkTo.sort()
-        );
-        expect(newState.timeline.linkTo.sort()).toEqual(
-          socTrendsGlobalLinkToState.timeline.linkTo.sort()
-        );
-        expect(newState.socTrends?.linkTo.sort()).toEqual(
-          socTrendsGlobalLinkToState.socTrends?.linkTo.sort()
-        );
-      });
-      test('socTrends, timeline, and global linked, do not update state', () => {
-        const inputState: InputsModel = allLinkToState;
-        const newState: InputsModel = addInputLink(
-          [InputsModelId.global, InputsModelId.socTrends],
-          inputState
-        );
-        expect(newState.global.linkTo.sort()).toEqual(inputState.global.linkTo.sort());
-        expect(newState.timeline.linkTo.sort()).toEqual(inputState.timeline.linkTo.sort());
-        expect(newState.socTrends?.linkTo.sort()).toEqual(inputState.socTrends?.linkTo.sort());
       });
     });
   });
 
   describe('#removeInputLink', () => {
     describe(`does not allow bad values`, () => {
-      test('More than 2 linkToIds passed', () => {
-        expect(() =>
-          removeInputLink(
-            [InputsModelId.global, InputsModelId.timeline, InputsModelId.socTrends],
-            state
-          )
-        ).toThrow('Only remove linkTo from 2 input states at a time');
-      });
       test('Less than 2 linkToIds passed', () => {
         expect(() => removeInputLink([InputsModelId.global], state)).toThrow(
           'Only remove linkTo from 2 input states at a time'
         );
       });
+
       test('Identical linkToIds passed', () => {
         expect(() => removeInputLink([InputsModelId.global, InputsModelId.global], state)).toThrow(
           'Input linkTo cannot remove link to itself'
         );
       });
-      test('Do not link timeline and socTrends', () => {
-        expect(() =>
-          removeInputLink([InputsModelId.timeline, InputsModelId.socTrends], state)
-        ).toThrow('Do not remove link socTrends to timeline. Only remove link socTrends to global');
-      });
     });
+
     describe(`linkToIds === ["global", "timeline"]`, () => {
       test('no inputs linked, do nothing', () => {
         const newState: InputsModel = removeInputLink(
@@ -617,17 +312,8 @@ describe('Inputs', () => {
         );
         expect(newState.global.linkTo).toEqual(emptyLinkToState.global.linkTo);
         expect(newState.timeline.linkTo).toEqual(emptyLinkToState.timeline.linkTo);
-        expect(newState.socTrends?.linkTo).toEqual(emptyLinkToState.socTrends?.linkTo);
       });
-      test('socTrends and global linked, do nothing', () => {
-        const newState: InputsModel = removeInputLink(
-          [InputsModelId.global, InputsModelId.timeline],
-          socTrendsGlobalLinkToState
-        );
-        expect(newState.global.linkTo).toEqual(socTrendsGlobalLinkToState.global.linkTo);
-        expect(newState.timeline.linkTo).toEqual(socTrendsGlobalLinkToState.timeline.linkTo);
-        expect(newState.socTrends?.linkTo).toEqual(socTrendsGlobalLinkToState.socTrends?.linkTo);
-      });
+
       test('timeline and global linked, remove link', () => {
         const newState: InputsModel = removeInputLink(
           [InputsModelId.global, InputsModelId.timeline],
@@ -635,54 +321,15 @@ describe('Inputs', () => {
         );
         expect(newState.global.linkTo).toEqual([]);
         expect(newState.timeline.linkTo).toEqual([]);
-        expect(newState.socTrends?.linkTo).toEqual([]);
       });
-      test('socTrends, timeline, and global linked, unlink timeline/global only', () => {
+
+      test('timeline and global linked, unlink timeline/global only', () => {
         const newState: InputsModel = removeInputLink(
           [InputsModelId.global, InputsModelId.timeline],
           allLinkToState
         );
-        expect(newState.global.linkTo).toEqual([InputsModelId.socTrends]);
-        expect(newState.timeline.linkTo).toEqual([]);
-        expect(newState.socTrends?.linkTo).toEqual([InputsModelId.global]);
-      });
-    });
-    describe(`linkToIds === ["global", "socTrends"]`, () => {
-      test('no inputs linked, do nothing', () => {
-        const newState: InputsModel = removeInputLink(
-          [InputsModelId.global, InputsModelId.socTrends],
-          emptyLinkToState
-        );
-        expect(newState.global.linkTo).toEqual(emptyLinkToState.global.linkTo);
-        expect(newState.timeline.linkTo).toEqual(emptyLinkToState.timeline.linkTo);
-        expect(newState.socTrends?.linkTo).toEqual(emptyLinkToState.socTrends?.linkTo);
-      });
-      test('timeline and global linked, do nothing', () => {
-        const newState: InputsModel = removeInputLink(
-          [InputsModelId.global, InputsModelId.socTrends],
-          timelineGlobalLinkToState
-        );
-        expect(newState.global.linkTo).toEqual(timelineGlobalLinkToState.global.linkTo);
-        expect(newState.timeline.linkTo).toEqual(timelineGlobalLinkToState.timeline.linkTo);
-        expect(newState.socTrends?.linkTo).toEqual(timelineGlobalLinkToState.socTrends?.linkTo);
-      });
-      test('socTrends and global linked, remove link', () => {
-        const newState: InputsModel = removeInputLink(
-          [InputsModelId.global, InputsModelId.socTrends],
-          socTrendsGlobalLinkToState
-        );
         expect(newState.global.linkTo).toEqual([]);
         expect(newState.timeline.linkTo).toEqual([]);
-        expect(newState.socTrends?.linkTo).toEqual([]);
-      });
-      test('socTrends, timeline, and global linked, unlink timeline/global only', () => {
-        const newState: InputsModel = removeInputLink(
-          [InputsModelId.global, InputsModelId.socTrends],
-          allLinkToState
-        );
-        expect(newState.global.linkTo).toEqual([InputsModelId.timeline]);
-        expect(newState.timeline.linkTo).toEqual([InputsModelId.global]);
-        expect(newState.socTrends?.linkTo).toEqual([]);
       });
     });
   });
@@ -716,7 +363,7 @@ describe('Inputs', () => {
         })
       ).toEqual({
         global: {
-          linkTo: [InputsModelId.timeline, InputsModelId.socTrends],
+          linkTo: [InputsModelId.timeline],
           policy: {
             duration: 300000,
             kind: 'manual',
@@ -742,7 +389,7 @@ describe('Inputs', () => {
           filters: [],
         },
         timeline: {
-          linkTo: [InputsModelId.global, InputsModelId.socTrends],
+          linkTo: [InputsModelId.global],
           policy: {
             duration: 300000,
             kind: 'manual',
@@ -759,7 +406,6 @@ describe('Inputs', () => {
           filters: [],
         },
         valueReport: state.valueReport,
-        socTrends: state.socTrends,
       });
     });
   });
