+++ x-pack/solutions/security/plugins/security_solution/public/common/store/inputs/model.ts
@@ -94,8 +94,6 @@ export interface InputsModel {
   global: InputsRange;
   timeline: InputsRange;
   valueReport: InputsRangeTimeOnly;
-  // TODO: remove ? when isSocTrendsEnabled feature flag is removed
-  socTrends?: InputsRangeTimeOnly;
 }
 export interface UrlInputsModelInputs {
   linkTo: InputsModelId[];
@@ -105,6 +103,4 @@ export interface UrlInputsModel {
   global: UrlInputsModelInputs;
   timeline: UrlInputsModelInputs;
   valueReport: UrlInputsModelInputs;
-  // TODO: remove ? when isSocTrendsEnabled feature flag is removed
-  socTrends?: UrlInputsModelInputs;
 }
