--- x-pack/solutions/security/plugins/security_solution/public/overview/components/detection_response/soc_trends/index.tsx
@@ -1,8 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
- */
-
-export { SocTrends } from './soc_trends';
