+++ x-pack/solutions/security/plugins/security_solution/public/reports/components/ai_value/utils.ts
@@ -4,10 +4,10 @@
  * 2.0; you may not use this file except in compliance with the Elastic License
  * 2.0.
  */
-import type { EuiThemeComputed } from '@elastic/eui';
-import * as i18n from './translations';
 
-import { getPercChange } from '../../../overview/components/detection_response/soc_trends/helpers';
+import type { EuiThemeComputed } from '@elastic/eui';
+import { getPercChange } from './helpers';
+import * as i18n from './translations';
 
 export const getPercentInfo = ({
   colors,
